+++ LibrarySystem/backend/src/main/java/com/pvt/jd2/project/back/domain/RoleImpl.java
@@ -15,7 +15,7 @@ import java.util.List;
  */
 public class RoleImpl implements Serializable, Role {
 
-    private Integer id;
+    private Long id;
     private String name;
     private List<Attribute> attributes;
     private List<User> users;
@@ -24,12 +24,12 @@ public class RoleImpl implements Serializable, Role {
     }
 
     @Override
-    public void setId(Integer integer) {
+    public void setId(Long integer) {
         this.id = id;
     }
 
     @Override
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
