+++ LibrarySystem/common/src/main/java/com/pvt/jd2/project/common/domain/BookSerial.java
@@ -10,9 +10,9 @@ import java.util.List;
  */
 public interface BookSerial {
 
-    void setId(Integer id);
+    void setId(Long id);
 
-    Integer getId();
+    Long getId();
 
     void setName(String name);
