+++ LibrarySystem/common/src/main/java/com/pvt/jd2/project/common/domain/Role.java
@@ -10,9 +10,9 @@ import java.util.List;
  */
 public interface Role {
 
-    void setId(Integer id);
+    void setId(Long id);
 
-    Integer getId();
+    Long getId();
 
     void setName(String name);
